refactor components and update tokentab
