+++ components/BalanceCircle.tsx
@@ -1,34 +1,13 @@
 import {View} from 'react-native';
-import React, {memo, useEffect, useState, Fragment} from 'react';
-import {useWallet, useExchangeRate} from '@hooks';
+import React, {memo, useEffect, useState} from 'react';
+import {useWallet} from '@hooks';
 import {Connection_Stats_Enum} from '../constants/Type';
 import CurrencyText from './CurrencyText';
 import Text from './Text';
 import {scale, verticalScale} from 'react-native-size-matters';
 import {AnimatedSegments} from './AnimatedSegments';
 import useTraceUpdates from '../src/hooks/useTraceUpdates';
-
-export const ToFiat = (props: {totalAmount: number}) => {
-  const {selectedCurrency, currencyRate} = useExchangeRate();
-
-  let calculatedAmount: number = props?.totalAmount * currencyRate;
-
-  let currency = selectedCurrency ? selectedCurrency.toUpperCase() : '';
-
-  useEffect(() => {
-    calculatedAmount = props?.totalAmount * currencyRate;
-  }, [selectedCurrency, currencyRate, props?.totalAmount, calculatedAmount]);
-
-  return (
-    <Fragment>
-      <CurrencyText
-        children={calculatedAmount}
-        currency={currency}
-        {...props}
-      />
-    </Fragment>
-  );
-};
+import {ToFiat} from '@utils';
 
 const BalanceCircle = memo(() => {
   const {
@@ -110,13 +89,25 @@ const BalanceCircle = memo(() => {
       ) : (
         <View style={{position: 'absolute'}}>
           <Text style={{textAlign: 'center'}}>Balance:</Text>
-          <View style={{flexDirection: 'row', alignItems: 'center', justifyContent: 'center'}}>
+          <View
+            style={{
+              flexDirection: 'row',
+              alignItems: 'center',
+              justifyContent: 'center',
+            }}>
             <CurrencyText
+              adjustsFontSizeToFit
               category="caption1"
               children={totalBalance.toFixed(8)}
             />
-            <Text marginHorizontal={3} category='body'>/</Text>
-            <ToFiat category="caption1" totalAmount={totalBalance} />
+            <Text adjustsFontSizeToFit marginHorizontal={3} category="body">
+              /
+            </Text>
+            <ToFiat
+              adjustsFontSizeToFit
+              category="caption1"
+              totalAmount={totalBalance}
+            />
           </View>
         </View>
       )}
