+++ components/index.tsx
@@ -4,4 +4,5 @@ export {default as Content} from './Content';
 export {default as BottomSheetOptions} from './BottomSheetOptions';
 export {default as OptionCard} from './OptionCard';
 export {default as Text} from './Text';
-export {default as LoadingComponent} from './LoadingComponent';
+export {default as BalanceCircle} from './BalanceCircle';
+export {default as CurrencyText} from './CurrencyText';
