+++ src/screens/main/partials/TokensTab.tsx
@@ -1,4 +1,4 @@
-import {Pressable, RefreshControl, ScrollView, View} from 'react-native';
+import {RefreshControl, ScrollView, View} from 'react-native';
 import React, {useCallback, useState, useEffect} from 'react';
 import {BalanceFragment} from '@constants/Type';
 import {useBottomSheet, useWallet} from '@hooks';
@@ -6,52 +6,15 @@ import {
   StyleService,
   useStyleSheet,
   useTheme,
-  Icon,
 } from '@tsejerome/ui-kitten-components';
-import {FiatRequest} from '@service';
 import {scale, verticalScale} from 'react-native-size-matters';
 
-import {
-  BalanceCard,
-  BottomSheetMenu,
-  Content,
-  Text,
-  LoadingComponent,
-} from '@components';
-import Fiat from './FiatPrices';
+import {BalanceCard, BottomSheetMenu, Content, Text} from '@components';
 
 const Tokens = () => {
   const [account, setAccount] = useState<BalanceFragment | undefined>(
     undefined,
   );
-  const {data, isError, isFetching, refetch, remove} = FiatRequest();
-
-  const [fiatPrices, setFiatPrices] = useState();
-  const [isLoading, setIsLoading] = useState(false);
-
-  // eslint-disable-next-line @typescript-eslint/no-shadow
-  const getFiatPrices = useCallback(async () => {
-    // eslint-disable-next-line @typescript-eslint/no-shadow
-    try {
-      remove();
-      const res = await refetch();
-      if (isFetching) {
-        setIsLoading(true);
-      } else if (!isFetching) {
-        setFiatPrices(res.data['nav-coin']);
-        setIsLoading(false);
-      }
-    } catch (e) {}
-  }, [isFetching, refetch, remove]);
-
-  useEffect(() => {
-    if (isFetching) {
-      setIsLoading(true);
-    } else if (!isFetching) {
-      setFiatPrices(data['nav-coin']);
-      setIsLoading(false);
-    }
-  }, [isFetching, data]);
 
   const {refreshWallet, tokens} = useWallet();
   const bottomSheet = useBottomSheet();
@@ -149,47 +112,6 @@ const Tokens = () => {
         <Content style={{paddingTop: verticalScale(24)}}>
           {tokensContent}
         </Content>
-        <View
-          style={{
-            paddingBottom: 30,
-            flexDirection: 'row',
-            justifyContent: 'space-between',
-            alignItems: 'center',
-            paddingRight: 16,
-          }}>
-          <Text>Exchange rates</Text>
-
-          <Pressable onPress={() => getFiatPrices()}>
-            <Icon
-              style={{
-                width: 15,
-                height: 20,
-              }}
-              pack="eva"
-              name="sync-outline"
-              fill={theme['color-basic-200']}
-            />
-          </Pressable>
-        </View>
-        {fiatPrices ? (
-          <View>
-            {Object.keys(fiatPrices).map((key, index) => {
-              return (
-                <View key={index}>
-                  <Fiat currency={key} value={fiatPrices[key]} />
-                </View>
-              );
-            })}
-          </View>
-        ) : (
-          <View>
-            <LoadingComponent
-              isLoading={isLoading}
-              message={undefined}
-              error={isError}
-            />
-          </View>
-        )}
       </ScrollView>
     </View>
   );
