+++ utils/index.tsx
@@ -1,0 +1,1 @@
+export {default as ToFiat} from './toFiat';
